Add files via upload
What is included in the PR?
included in the PR are three files. the main one is the assignment04 file (the old-assignment04 is just one i was working on and did not use), as well as the venue code that was given for the assignment.

Explain the changes being made.
Week04 assignments will be pushed to the main branch. I have not created a test script because I wanted to see what was needed to be tested and will create a test script once I know..

What difficulties you ran into, and the resulting architectural choices you made, while doing the assignment?
My main difficulty was understanding the assignment. I felt that the more I read it, the more it confused me. I did not know if the supplemental file (venue.go) was meant for use to use or what we needed the assignment04 to point to. Also, I did understand how to make a pointer receiver point to an interface. I kept running into error codes when using *venue.

While surprises did you find while doing the assignment?
Interfaces is not a hard concept to understand. Well I should say the introduction of it. I guess I need to understand more of how we are to use this in a real world situation. also, i want to understand how you can use it with pointer receivers. However, I am still confused as to how testing works and what needs to be done to get it working correctly.